chore(connector): Cleanup base models and views
