+++ superset/common/query_context_factory.py
@@ -29,7 +29,7 @@ from superset.models.slice import Slice
 from superset.utils.core import DatasourceDict, DatasourceType, is_adhoc_column
 
 if TYPE_CHECKING:
-    from superset.connectors.base.models import BaseDatasource
+    from superset.connectors.sqla.models import BaseDatasource
 
 config = app.config
