+++ superset/datasets/commands/importers/v0.py
@@ -26,8 +26,12 @@ from sqlalchemy.orm.session import make_transient
 from superset import db
 from superset.commands.base import BaseCommand
 from superset.commands.importers.exceptions import IncorrectVersionError
-from superset.connectors.base.models import BaseColumn, BaseDatasource, BaseMetric
-from superset.connectors.sqla.models import SqlaTable, SqlMetric, TableColumn
+from superset.connectors.sqla.models import (
+    BaseDatasource,
+    SqlaTable,
+    SqlMetric,
+    TableColumn,
+)
 from superset.databases.commands.exceptions import DatabaseNotFoundError
 from superset.datasets.commands.exceptions import DatasetInvalidError
 from superset.models.core import Database
@@ -102,14 +106,8 @@ def lookup_sqla_metric(session: Session, metric: SqlMetric) -> SqlMetric:
     )
 
 
-def import_metric(session: Session, metric: BaseMetric) -> BaseMetric:
-    if isinstance(metric, SqlMetric):
-        lookup_metric = lookup_sqla_metric
-    else:
-        raise Exception(  # pylint: disable=broad-exception-raised
-            f"Invalid metric type: {metric}"
-        )
-    return import_simple_obj(session, metric, lookup_metric)
+def import_metric(session: Session, metric: SqlMetric) -> SqlMetric:
+    return import_simple_obj(session, metric, lookup_sqla_metric)
 
 
 def lookup_sqla_column(session: Session, column: TableColumn) -> TableColumn:
@@ -123,14 +121,8 @@ def lookup_sqla_column(session: Session, column: TableColumn) -> TableColumn:
     )
 
 
-def import_column(session: Session, column: BaseColumn) -> BaseColumn:
-    if isinstance(column, TableColumn):
-        lookup_column = lookup_sqla_column
-    else:
-        raise Exception(  # pylint: disable=broad-exception-raised
-            f"Invalid column type: {column}"
-        )
-    return import_simple_obj(session, column, lookup_column)
+def import_column(session: Session, column: TableColumn) -> TableColumn:
+    return import_simple_obj(session, column, lookup_sqla_column)
 
 
 def import_datasource(  # pylint: disable=too-many-arguments
