+++ tests/integration_tests/base_tests.py
@@ -36,8 +36,7 @@ from sqlalchemy.dialects.mysql import dialect
 from tests.integration_tests.test_app import app, login
 from superset.sql_parse import CtasMethod
 from superset import db, security_manager
-from superset.connectors.base.models import BaseDatasource
-from superset.connectors.sqla.models import SqlaTable
+from superset.connectors.sqla.models import BaseDatasource, SqlaTable
 from superset.models import core as models
 from superset.models.slice import Slice
 from superset.models.core import Database
